put all cryptographic operations to `Crypto`
