+++ interpreter/src/main/scala/fssi/interpreter/util/package.scala
@@ -1,5 +1,5 @@
 package fssi.interpreter
 
 package object util {
-  object sha3 extends SHA3
+  object crypto extends Crypto
 }
